Add tenax.linalg module with svd, qr, eigh and block-sparse CTM projector
Create tenax.linalg as the canonical home for tensor decompositions:
- Move truncated_svd → svd and qr_decompose → qr from contractor.py
- Add new eigh() for Hermitian eigendecomposition with block-sparse dispatch
- Backward-compatible re-exports in contractor.py

Add _eigh_projector_symmetric in _ctm_tensor.py for per-sector density
matrix eigendecomposition in CTM projector computation, avoiding dense
round-trip for SymmetricTensor inputs. Falls back to dense path when
blocks contain JAX tracers (AD) or when grown corners have no blocks
(e.g., FermionParity early iterations).
