+++ src/tenax/algorithms/_ctm_tensor.py
@@ -76,6 +76,122 @@ def _fuse_pair_by_label(
     return fuse_indices(T, axis_a, axis_b, fused_label, fused_flow)
 
 
+def _eigh_projector_symmetric(
+    C1g: SymmetricTensor,
+    C4g: SymmetricTensor,
+    chi: int,
+) -> SymmetricTensor:
+    r"""Block-sparse projector via per-sector density matrix eigh.
+
+    For each charge sector *q* of the ``fused`` leg, accumulates
+    :math:`\rho_q = M_1 M_1^\dagger + M_4 M_4^\dagger` where :math:`M_i`
+    is the sector's dense block of C_ig, then eigendecomposes :math:`\rho_q`.
+    Eigenvalues are merged across sectors and globally truncated to *chi*.
+
+    The result is wrapped via ``from_dense`` with trivial ``chi_new`` charges
+    so that downstream label-based contractions remain charge-consistent with
+    the existing CTM environment structure.
+
+    Args:
+        C1g: Grown corner SymmetricTensor ``(fused, col1)``.
+        C4g: Grown corner SymmetricTensor ``(fused, col2)``.
+        chi: Target bond dimension.
+
+    Returns:
+        Projector ``P`` with labels ``(fused, chi_new)``, flows ``(IN, OUT)``.
+    """
+    fused_pos = C1g.labels().index("fused")
+    col_pos = 1 - fused_pos  # the other leg
+    fused_idx = C1g.indices[fused_pos]
+    sym = fused_idx.symmetry
+    fused_total = fused_idx.dim
+
+    # Group blocks by fused charge for each corner
+    def _group_by_fused(Cg: SymmetricTensor) -> dict[int, list[tuple[int, jax.Array]]]:
+        """Map fused_charge -> list of (col_charge, block)."""
+        grouped: dict[int, list[tuple[int, jax.Array]]] = {}
+        for key, block in Cg.blocks.items():
+            fq = int(key[fused_pos])
+            cq = int(key[col_pos])
+            grouped.setdefault(fq, []).append((cq, block))
+        return grouped
+
+    c1_groups = _group_by_fused(C1g)
+    c4_groups = _group_by_fused(C4g)
+
+    all_fused_charges = sorted(set(c1_groups.keys()) | set(c4_groups.keys()))
+
+    # Per-sector eigh results: q -> (eigvecs, eigvals, fused_dim, row_offset)
+    sector_results: dict[int, tuple[jax.Array, jax.Array, int, int]] = {}
+
+    # Build a map from fused charge to its row indices in the dense fused index
+    charges_arr = np.asarray(fused_idx.charges)
+    charge_rows: dict[int, np.ndarray] = {}
+    for fq in all_fused_charges:
+        charge_rows[fq] = np.where(charges_arr == fq)[0]
+
+    for fq in all_fused_charges:
+        fused_dim = int(len(charge_rows.get(fq, [])))
+        if fused_dim == 0:
+            continue
+
+        # Accumulate rho = sum of M @ M^dagger for both corners
+        rho = jnp.zeros((fused_dim, fused_dim), dtype=C1g.dtype)
+
+        for entries in [c1_groups.get(fq, []), c4_groups.get(fq, [])]:
+            for _cq, block in entries:
+                if fused_pos == 0:
+                    M = block.reshape(fused_dim, -1)
+                else:
+                    M = block.reshape(-1, fused_dim).T
+                rho = rho + M @ M.conj().T
+
+        rho = 0.5 * (rho + rho.conj().T)
+        eigvals, eigvecs = jnp.linalg.eigh(rho)
+        sector_results[fq] = (eigvecs, eigvals, fused_dim, charge_rows[fq])
+
+    # Global truncation: merge eigenvalues, keep top-chi
+    all_eig_pairs: list[tuple[float, int, int]] = []  # (value, fused_charge, index)
+    for fq, (_, eigvals, _, _) in sector_results.items():
+        for i, val in enumerate(np.array(eigvals)):
+            all_eig_pairs.append((float(val), fq, i))
+
+    # Sort descending by eigenvalue, then descending by index to match
+    # the dense convention of taking eigvecs[:, -k:] (highest indices first
+    # among degenerate eigenvalues).
+    all_eig_pairs.sort(key=lambda x: (-x[0], -x[2]))
+    n_keep = min(chi, len(all_eig_pairs))
+
+    # Count per-sector keeps
+    sector_keep: dict[int, list[int]] = {}
+    for _, fq, idx in all_eig_pairs[:n_keep]:
+        sector_keep.setdefault(fq, []).append(idx)
+
+    # Assemble dense projector from per-sector eigenvectors, then wrap via
+    # from_dense with trivial chi_new charges.  Only the identity-charge
+    # sector survives conservation filtering, matching the dense path.
+    P_dense = jnp.zeros((fused_total, n_keep), dtype=C1g.dtype)
+    col = 0
+    for fq in sorted(sector_keep.keys()):
+        keep_indices = sorted(sector_keep[fq], reverse=True)
+        n_q = len(keep_indices)
+
+        eigvecs, _, fused_dim, row_idx = sector_results[fq]
+        V_q = eigvecs[:, keep_indices]  # (fused_dim, n_q)
+
+        P_dense = P_dense.at[row_idx, col : col + n_q].set(V_q)
+        col += n_q
+
+    P_dense = jax.lax.stop_gradient(P_dense)
+
+    chi_new_idx = TensorIndex(
+        sym, np.zeros(n_keep, dtype=np.int32), OUT, label="chi_new"
+    )
+    return SymmetricTensor.from_dense(
+        P_dense, (fused_idx, chi_new_idx), tol=float("inf")
+    )
+
+
 def _compute_projector_tensor(
     C1g: Tensor,
     C4g: Tensor,
@@ -83,9 +199,12 @@ def _compute_projector_tensor(
 ) -> Tensor:
     r"""Compute isometric projector P as a Tensor via eigh of the density matrix.
 
-    Forms :math:`\rho = C_{1g} C_{1g}^\dagger + C_{4g} C_{4g}^\dagger`
-    in dense space, eigendecomposes via ``eigh``, then wraps the top-k
-    eigenvectors as a Tensor with indices matching C1g's ``fused`` leg.
+    Forms :math:`\rho = C_{1g} C_{1g}^\dagger + C_{4g} C_{4g}^\dagger`,
+    eigendecomposes, then wraps the top-k eigenvectors as a Tensor with
+    indices matching C1g's ``fused`` leg.
+
+    For SymmetricTensor inputs, uses block-sparse eigh (per charge sector)
+    to avoid dense round-trip.
 
     Args:
         C1g: Grown corner with labels ``(fused, <col1>)``.
@@ -96,6 +215,14 @@ def _compute_projector_tensor(
         Projector ``P`` with labels ``(fused, chi_new)``,
         flows ``(IN, OUT)``.  Wrapped in ``stop_gradient``.
     """
+    # Use block-sparse path for SymmetricTensor unless blocks contain
+    # JAX tracers (during AD), in which case fall back to dense path
+    # since eigenvalue sorting requires concrete values.
+    if isinstance(C1g, SymmetricTensor) and isinstance(C4g, SymmetricTensor):
+        has_tracers = any(isinstance(b, jax.core.Tracer) for b in C1g.blocks.values())
+        if not has_tracers and C1g.blocks and C4g.blocks:
+            return _eigh_projector_symmetric(C1g, C4g, chi)
+
     C1g_dense = C1g.todense()
     C4g_dense = C4g.todense()
